Factored the test area code out of the main maps file.
